refactor: move getAddressPrefix out of `validators.ts` into `index.ts`
